+++ src/middlewar/redirect-js.js
@@ -8,9 +8,9 @@
 
 export default function(req, res, next) {
     if (/\/@cdnhost[\s\S]+@version/g.test(req.path)) {
-        const filePath = req.path.replace(/\/@cdnhost[\s\S]+@version/g, '');
-        res.redirect(filePath);
-        return;
+        const filePath = req.path.replace(/\/@cdnhost[\s\S]+@version/g, '')
+        res.redirect(filePath)
+        return
     }
-    next();
+    next()
 }