+++ src/webpack/module.js
@@ -2,11 +2,6 @@ import Utils from '../utils/'
 import getRules from './getRules'
 export default function() {
     return {
-        rules: getRules(Utils.getUserConfig),
-        //不在扫正则所匹配的模块的依赖
-        noParse: function(content) {
-            const pass = /(lodash|zepto|jquery)\.js/.test(content)
-            return pass
-        }
+        rules: getRules(Utils.getUserConfig)
     }
 }