+++ src/webpack/plugins.js
@@ -2,29 +2,32 @@
 import path from 'path'
 import webpack from 'webpack'
 import Utils from '../utils'
-import HtmlWebpackPlugin from '../plugins/html-webpack-plugin'
+import HtmlWebpackPlugin from 'html-webpack-plugin'
 import htmlInlineSourceLoaders from '../plugins/html-inline-source-loaders'
-import HtmlWebpackReplaceHost from 'html-webpack-replace-host'
+// import HtmlWebpackReplaceHost from 'html-webpack-replace-host'
 import HtmlWebpackInlineSourcePlugin from 'webpack-plugin-inline-source'
 
 const tagVersion = Utils.getUserConfig.version
 
 function setHtmlPlugin(file, env) {
     const isOnline = ['tag', 'productionNoTag', 'production-build', 'production', 'gray'].indexOf(env) !== -1
+    const htmlOutputDir = path.resolve(process.cwd(), './deploy', isOnline ? './html/build' : './build')
+    const cwdDir = path.resolve(process.cwd(), htmlOutputDir || './dist')
     return new HtmlWebpackPlugin({
         version: tagVersion,
-        inject: true,
-        hash: true,
+        inject: 'body', // true | 'head' | 'body' ， 主js插入位置(head \ body)
+        hash: true, // true | false, 插入的js文件src后面增加?[hash]值，每次构建唯一
         minify: {
             collapseWhitespace: true,
             removeComments: true
         },
+        // 配置加入的 <meta/>
         meta: {
             env: env,
             'update-time': { 'update-time': new Date().toLocaleString() }
         },
-        path: path.resolve(process.cwd(), './deploy', isOnline ? './html/build' : './build'),
-        filename: file,
+        path: htmlOutputDir,
+        filename: path.resolve(cwdDir, file),
         template: path.resolve(file)
     })
 }
@@ -46,26 +49,30 @@ function getPlugins({ htmlEntry, env = 'daily', cdnhost }) {
 
     if (env == 'local' || env == 'daily' || env == 'production-build') {
         config.push(new webpack.ProgressPlugin({ percentBy: 'entries' }))
-        // config.push(new LazyPathPlugin({ version: tagVersion, jsHost, env }))
+    }
+
+    // config.push(new LazyPathPlugin({ version: tagVersion, jsHost, env }))
+    if (env !== 'local') {
         config.push(setHtmlPlugin(htmlEntry, env))
 
         // console.log(jsPath)
-        config.push(
-            new HtmlWebpackReplaceHost({
-                replaceString: env == 'local' || env == 'daily' ? '' : jsPath
-            })
-        )
-
+        // config.push(
+        //     new HtmlWebpackReplaceHost({
+        //         replaceString: env == 'local' || env == 'daily' ? '' : jsPath
+        //     })
+        // )
         config.push(
             new HtmlWebpackInlineSourcePlugin({
+                /* html-webpack-plugin, v5.5 版本
+                 * 不传入getHooks，它不执行相关hooks
+                 */
+                htmlWebpackPluginHooks: HtmlWebpackPlugin.getHooks,
+                // 对inline的js代码二次加工
                 loaders: htmlInlineSourceLoaders
             })
         )
-    } else {
-        // config.push(new webpack.NamedModulesPlugin())
-        // config.push(new FriendlyErrorsPlugin())
     }
-    // console.log(config);
+
     return config
 }