+++ src/webpack/webpack.pro.config.js
@@ -2,6 +2,7 @@ import path from 'path'
 import Utils from '../utils/'
 import pluginsConfig from './plugins'
 import performance from './performance'
+import optimization from './optimization'
 import { getExternals } from './externals'
 import { merge } from 'webpack-merge'
 import webpackBaseConfig from './webpack.base.config'
@@ -19,7 +20,7 @@ export default (parentDirPath, option = {}) => (outputPath, buildConfig) => {
         webpackBaseConfig(parentDirPath),
         {
             mode: 'production',
-            performance: performance,
+            // performance: performance,
             entry: {},
             output: {
                 path: path.resolve(cwdPath, output),
@@ -27,6 +28,8 @@ export default (parentDirPath, option = {}) => (outputPath, buildConfig) => {
                 // chunkFilename: _filename
                 publicPath: isOnline ? jsPath + '/' : `./${Utils.getUserConfig.version}/`
             },
+            //打包多文件和公共模块配置, dev环境下没有使用html-webpack-plugin，无法自动引入分离的包
+            optimization: optimization, // webpack5下部分被废弃
             resolve: {
                 modules: [path.join(parentDirPath, 'node_modules')]
             },